Refactor historical metadata collection
